refactor(backend): extract websocket layer and document modularization plan
